Move generic function to util file
